requirements.txt
torch --index-url https://download.pytorch.org/whl/cu118
transformers
fastapi 
uvicorn
python-dotenv